# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  imageUrls: string[];
  /** @nullable */
  videoUrl?: string | null;
  inStock: boolean;
  featured: boolean;
  sizes?: string[];
  materials?: string[];
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  leadTime?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  tags?: string[];
}

export interface ProductInput {
  name: string;
  slug?: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  imageUrls: string[];
  /** @nullable */
  videoUrl?: string | null;
  inStock?: boolean;
  featured?: boolean;
  sizes?: string[];
  materials?: string[];
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  leadTime?: string | null;
  tags?: string[];
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  category?: string;
  imageUrls?: string[];
  /** @nullable */
  videoUrl?: string | null;
  inStock?: boolean;
  featured?: boolean;
  sizes?: string[];
  materials?: string[];
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  leadTime?: string | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
  /** @nullable */
  selectedSize: string | null;
}

export interface Cart {
  id: number;
  sessionId: string;
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
  /** @nullable */
  selectedSize?: string | null;
}

export interface CartItemUpdate {
  quantity: number;
  sessionId?: string;
}

export interface ShippingAddress {
  street: string;
  city: string;
  state: string;
  zip: string;
  country: string;
}

export interface OrderInput {
  sessionId: string;
  customerName: string;
  email: string;
  phone: string;
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
}

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone?: string;
  shippingAddress?: ShippingAddress;
  items: CartItem[];
  subtotal?: number;
  shipping?: number;
  total: number;
  status: string;
  paymentMethod?: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface FeaturedCollection {
  heroProduct: Product;
  featuredProducts: Product[];
  newArrivals: Product[];
}

export type CatalogSummaryPriceRange = {
  min: number;
  max: number;
};

export type CatalogSummaryCategoryBreakdownItem = {
  category: string;
  count: number;
};

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  priceRange: CatalogSummaryPriceRange;
  categoryBreakdown?: CatalogSummaryCategoryBreakdownItem[];
}

export type ListProductsParams = {
category?: string;
featured?: boolean;
search?: string;
};

export type GetCartParams = {
session_id: string;
};