# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Resin Arts Studio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
  - name: catalog
    description: Catalog summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /catalog/featured:
    get:
      operationId: getFeaturedCollection
      tags: [catalog]
      summary: Get featured products collection
      responses:
        "200":
          description: Featured collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturedCollection"

  /catalog/summary:
    get:
      operationId: getCatalogSummary
      tags: [catalog]
      summary: Get catalog summary stats
      responses:
        "200":
          description: Catalog summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, name, description, price, category, imageUrls, inStock, featured, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        dimensions:
          type: ["string", "null"]
        leadTime:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string

    ProductInput:
      type: object
      required: [name, description, price, category, imageUrls]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        dimensions:
          type: ["string", "null"]
        leadTime:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        videoUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        dimensions:
          type: ["string", "null"]
        leadTime:
          type: ["string", "null"]

    Category:
      type: object
      required: [id, name, slug, productCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer

    Cart:
      type: object
      required: [id, sessionId, items, subtotal, itemCount]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        itemCount:
          type: integer

    CartItem:
      type: object
      required: [id, productId, productName, productImage, price, quantity, selectedSize]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        price:
          type: number
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]

    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
        sessionId:
          type: string

    OrderInput:
      type: object
      required: [sessionId, customerName, email, phone, shippingAddress, paymentMethod]
      properties:
        sessionId:
          type: string
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]

    Order:
      type: object
      required: [id, customerName, email, total, status, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    ShippingAddress:
      type: object
      required: [street, city, state, zip, country]
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string

    FeaturedCollection:
      type: object
      required: [heroProduct, featuredProducts, newArrivals]
      properties:
        heroProduct:
          $ref: "#/components/schemas/Product"
        featuredProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        newArrivals:
          type: array
          items:
            $ref: "#/components/schemas/Product"

    CatalogSummary:
      type: object
      required: [totalProducts, totalCategories, priceRange]
      properties:
        totalProducts:
          type: integer
        totalCategories:
          type: integer
        priceRange:
          type: object
          required: [min, max]
          properties:
            min:
              type: number
            max:
              type: number
        categoryBreakdown:
          type: array
          items:
            type: object
            required: [category, count]
            properties:
              category:
                type: string
              count:
                type: integer