# lib/api-zod/src/generated/types/cart.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface Cart {
  id: number;
  sessionId: string;
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}