# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogSummaryCategoryBreakdownItem } from './catalogSummaryCategoryBreakdownItem';
import type { CatalogSummaryPriceRange } from './catalogSummaryPriceRange';

export interface CatalogSummary {
  totalProducts: number;
  totalCategories: number;
  priceRange: CatalogSummaryPriceRange;
  categoryBreakdown?: CatalogSummaryCategoryBreakdownItem[];
}