# lib/api-zod/src/generated/types/featuredCollection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface FeaturedCollection {
  heroProduct: Product;
  featuredProducts: Product[];
  newArrivals: Product[];
}