# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';
import type { ShippingAddress } from './shippingAddress';

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone?: string;
  shippingAddress?: ShippingAddress;
  items: CartItem[];
  subtotal?: number;
  shipping?: number;
  total: number;
  status: string;
  paymentMethod?: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}