# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ShippingAddress } from './shippingAddress';

export interface OrderInput {
  sessionId: string;
  customerName: string;
  email: string;
  phone: string;
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
}