# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug?: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  imageUrls: string[];
  /** @nullable */
  videoUrl?: string | null;
  inStock?: boolean;
  featured?: boolean;
  sizes?: string[];
  materials?: string[];
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  leadTime?: string | null;
  tags?: string[];
}