# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Resin Arts Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingAddress {
  street: string;
  city: string;
  state: string;
  zip: string;
  country: string;
}